Adds new feature test file for new reviews
