+++ app/views/books/index.html.erb
@@ -9,7 +9,7 @@
 <%= "#{book.title} #{book.avg_rating}" %>
 <% end %>
 
-<%= link_to 'New Book', new_book_path %>
+<%= link_to 'Add New Book', new_book_path %>
 <% @books.each do |book| %>
 <h2><%= link_to "#{book.title}", book_path(book)%></h2>
 <p>Pages: <%= book.pages %></p>
